Use new struct variant feature
Struct variants are now available (behind a feature gate), which is just
as well since the compiler started complaining that all exported types
have to be pub. The good thing is that StyleStruct is now completely
unnecessary, and we can use and pattern-match on the Styled style as we
can all the others, which simplifies the code.