Make ANSIStrings Cow
This commit makes the ANSIString struct able to take not only string slices, but also owned strings, with a copy-on-write mechanism.

This has the glorious effect of reducing the number of String allocations in the programs that use it: it will no longer be necessary to return an owned string, but instead an ANSIString that holds the relevant data, no matter where it came from.

There’s also now a From impl, for convenience. The new method on ANSIString has been removed, as I’m pretty sure it was never meant to be public in the first place…