Fix build fail with i686 toolchain
The problem appeared to be the calling convention of of the kernel32 DLL
function and u64 being used even with i686 (32-bit) toolchain. On
Windows, a `DWORD` is always 32-bit (u32) even on a 64-bit system
(x86_64). The `GetStdHandle` declaration was using a u64.

Additionally, the `system` calling convention needed to be added because
in 64-bit environments (x86_64), the calling convention is "C" but in
32-bit environments the calling convention is "stdcall". The `system`
calling convention automatically handles the difference depending on the
environment/toolchain.